requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-chess>=1.9.0
pydantic>=2.0.0